pyproject.toml
[project]
name = "gpu-micro-bench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.8",
]
